pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Connection App"
authors = ["Cristi P <cristian.alexandru86@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
gunicorn = "^20.1.0"
tweepy = "^4.5.0"
Flask-PyMongo = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
safety = "^1.10.3"
pydocstyle = "^6.1.1"
flake8 = "^4.0.1"
flake8-cognitive-complexity = "^0.1.0"
black = "^22.1.0"
pre-commit = "^2.17.0"
bandit = "^1.7.2"
isort = "^5.10.1"
pytest-cov = "^3.0.0"
mypy = "^0.931"
types-requests = "^2.27.9"
pytest-coverage = "^0.0"
requests-mock = "^1.9.3"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = "manage:run_server"
